Major refactoring for long-term maintenance.
