+++ pom.xml
@@ -6,7 +6,7 @@
 	<modelVersion>4.0.0</modelVersion>
 	<groupId>com.logpresso</groupId>
 	<artifactId>log4j2-scanner</artifactId>
-	<version>2.2.0</version>
+	<version>2.2.1</version>
 	<packaging>jar</packaging>
 	<name>Logpresso Log4j2 Scanner</name>
