+++ src/main/java/com/logpresso/scanner/utils/DummyInputStream.java
@@ -1,4 +1,4 @@
-package com.logpresso.scanner;
+package com.logpresso.scanner.utils;
 
 import java.io.IOException;
 import java.io.InputStream;
