+++ src/main/java/com/logpresso/scanner/utils/FileUtils.java
@@ -1,0 +1,67 @@
+package com.logpresso.scanner.utils;
+
+import java.io.Closeable;
+import java.io.File;
+import java.io.FileInputStream;
+import java.io.FileOutputStream;
+import java.io.RandomAccessFile;
+import java.nio.channels.FileLock;
+
+public class FileUtils {
+	private FileUtils() {
+	}
+
+	public static boolean isLocked(File f) {
+		RandomAccessFile raf = null;
+		try {
+			raf = new RandomAccessFile(f, "rw");
+			FileLock lock = raf.getChannel().lock();
+			lock.release();
+			return false;
+		} catch (Throwable t) {
+			return true;
+		} finally {
+			IoUtils.ensureClose(raf);
+		}
+	}
+
+	public static boolean truncate(File f) {
+		RandomAccessFile raf = null;
+		try {
+			raf = new RandomAccessFile(f, "rw");
+			raf.setLength(0);
+			return true;
+		} catch (Throwable t) {
+			return false;
+		} finally {
+			IoUtils.ensureClose(raf);
+		}
+	}
+
+	public static boolean copyAsIs(File srcFile, File dstFile) {
+		FileInputStream is = null;
+		FileOutputStream os = null;
+
+		try {
+			is = new FileInputStream(srcFile);
+			os = new FileOutputStream(dstFile);
+
+			byte[] buf = new byte[32768];
+			while (true) {
+				int len = is.read(buf);
+				if (len < 0)
+					break;
+
+				os.write(buf, 0, len);
+			}
+
+			return true;
+		} catch (Throwable t) {
+			System.out.println("Error: Cannot copy file " + srcFile.getAbsolutePath() + " - " + t.getMessage());
+			return false;
+		} finally {
+			IoUtils.ensureClose(is);
+			IoUtils.ensureClose(os);
+		}
+	}
+}
