+++ src/main/java/com/logpresso/scanner/utils/StringUtils.java
@@ -1,0 +1,28 @@
+package com.logpresso.scanner.utils;
+
+import java.util.List;
+
+public class StringUtils {
+	private StringUtils() {
+	}
+
+	public static String toString(List<String> pathChain) {
+		if (pathChain == null)
+			return "";
+
+		return join(pathChain, " > ");
+	}
+
+	public static String join(List<String> tokens, String separator) {
+		StringBuilder sb = new StringBuilder();
+		int i = 0;
+
+		for (String path : tokens) {
+			if (i++ != 0)
+				sb.append(separator);
+			sb.append(path);
+		}
+
+		return sb.toString();
+	}
+}
